pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "shogi-vision"
version = "0.1.0"
description = "Convert shogi board images to SFEN notation using computer vision"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "opencv-python>=4.8",
    "numpy>=1.24",
    "Pillow>=10.0",
    "pytesseract>=0.3.10",
]

[project.optional-dependencies]
ml = ["torch>=2.0", "torchvision>=0.15"]
dev = ["pytest>=7.0", "pytest-cov>=4.0"]

[project.scripts]
shogi-vision = "shogi_vision.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
